requirements.txt
# pip install -r requirements.txt
numpy
sense_hat
python-dotenv
Pillow
cloudant